+++ php/Turkce.php
@@ -8,7 +8,7 @@
  * License: MIT
  * https://github.com/SiisOfficial/turkish-suffixes
  *
- * Version: 1.1
+ * Version: 1.2
  */
 class Turkce {
 
@@ -17,19 +17,12 @@ class Turkce {
     | Turkish Case Suffixes Class
     |--------------------------------------------------------------------------
     |
-    |   Because of the PHP's UTF-8 case-insensitive bug, this Class may not be
-    |   work correctly for uppercase names. We'll update it later.
-    |
-    |   Should work with PHP >=4.0.6, but we recommend PHP >=5.
+    |   Should work with PHP >=5.
     |   Not tested with PHP 7.
     |
     |   --------
     |
-    |   PHP'deki UTF-8 karakterlerdeki küçük/büyük harf duyarlılığı ile ilgili sıkıntı
-    |   yüzünden büyük harfli tüm isimlere ekleri doğru getirmeyebilir. Yakında
-    |   güncelleyeceğiz.
-    |
-    |   PHP 4.0.6 ve sonrasında çalışıyor olmalı; fakat PHP 5 ve yukarısı önerilir.
+    |   PHP 5 ve sonrasında çalışıyor olmalı.
     |   PHP 7 ile test edilmedi.
     |
     */
@@ -38,10 +31,16 @@ class Turkce {
      * Accusative case
      *
      * @param  string $noun
+     * @param bool $fake (integer/string if it's a fake noun): this is for pronounce usage: this is for pronounce usage
      *
      * @return string
      */
-    public static function belirtmeHali($noun) {
+    public static function belirtmeHali($noun, $fake = false) {
+        $noun = trim($noun);
+
+        //  Check the ending if it's a number
+        if(preg_match('/\d$/', $noun)) return self::processForNumber(__FUNCTION__, $noun);
+
         $last_letter      = mb_substr($noun, -1, 1, 'utf-8');
         $last_3_letter    = mb_substr($noun, -3, 3, 'utf-8');
         $nouns            = preg_split('/ /', $noun);
@@ -58,26 +57,32 @@ class Turkce {
 
         if(count($sh[0]) == 0) return ($noun . "'" . $blending . $suffix);
 
-        if(preg_match('/(*UTF8)[eiEİ]/', $sh[0][count($sh[0]) - 1])) $suffix = "i"; //  probably will be removed in next version.
         if(preg_match('/(*UTF8)[öüÖÜ]/', $sh[0][count($sh[0]) - 1])) $suffix = "ü";
         if(preg_match('/(*UTF8)[aıAI]/', $sh[0][count($sh[0]) - 1])) $suffix = "ı";
         if(preg_match('/(*UTF8)[ouOU]/', $sh[0][count($sh[0]) - 1])) $suffix = "u";
 
+        if($fake !== false) $noun = $fake;
         return ($noun . "'" . $blending . $suffix);
     }
 
-    public static function accusativeCase($noun) {
-        return self::belirtmeHali($noun);
+    public static function accusativeCase($noun, $fake = false) {
+        return self::belirtmeHali($noun, $fake);
     }
 
     /**
      * Dative case
      *
      * @param  string $noun
+     * @param bool $fake (integer/string if it's a fake noun): this is for pronounce usage
      *
      * @return string
      */
-    public static function yonelmeHali($noun) {
+    public static function yonelmeHali($noun, $fake = false) {
+        $noun = trim($noun);
+
+        //  Check the ending if it's a number
+        if(preg_match('/\d$/', $noun)) return self::processForNumber(__FUNCTION__, $noun);
+
         $last_letter      = mb_substr($noun, -1, 1, 'utf-8');
         $last_3_letter    = mb_substr($noun, -3, 3, 'utf-8');
         $nouns            = preg_split('/ /', $noun);
@@ -94,24 +99,30 @@ class Turkce {
 
         if(count($sh[0]) == 0) return ($noun . "'" . $blending . $suffix);
 
-        if(preg_match('/(*UTF8)[öüeiÖÜEİ]/', $sh[0][count($sh[0]) - 1])) $suffix = "e"; //  probably will be removed in next version.
         if(preg_match('/(*UTF8)[ouaıOUAI]/', $sh[0][count($sh[0]) - 1])) $suffix = "a";
 
+        if($fake !== false) $noun = $fake;
         return ($noun . "'" . $blending . $suffix);
     }
 
-    public static function dativeCase($noun) {
-        return self::yonelmeHali($noun);
+    public static function dativeCase($noun, $fake = false) {
+        return self::yonelmeHali($noun, $fake);
     }
 
     /**
      * Locative case
      *
      * @param  string $noun
+     * @param bool $fake (integer/string if it's a fake noun): this is for pronounce usage
      *
      * @return string
      */
-    public static function bulunmaHali($noun) {
+    public static function bulunmaHali($noun, $fake = false) {
+        $noun = trim($noun);
+
+        //  Check the ending if it's a number
+        if(preg_match('/\d$/', $noun)) return self::processForNumber(__FUNCTION__, $noun);
+
         $last_letter      = mb_substr($noun, -1, 1, 'utf-8');
         $last_3_letter    = mb_substr($noun, -3, 3, 'utf-8');
         $last_2_letter    = mb_substr($noun, -2, 2, 'utf-8');
@@ -137,21 +148,28 @@ class Turkce {
         if(preg_match('/(*UTF8)[ouaıOUAI]/', $sh[0][count($sh[0]) - 1])) $suffix_ = "a";
         if(preg_match('/(*UTF8)[öüeiÖÜEİ]/', $sh[0][count($sh[0]) - 1])) $suffix_ = "e";
 
+        if($fake !== false) $noun = $fake;
         return ($noun . "'" . $blending . $suffix . $suffix_);
     }
 
-    public static function locativeCase($noun) {
-        return self::bulunmaHali($noun);
+    public static function locativeCase($noun, $fake = false) {
+        return self::bulunmaHali($noun, $fake);
     }
 
     /**
      * Ablative case
      *
      * @param  string $noun
+     * @param bool $fake (integer/string if it's a fake noun): this is for pronounce usage
      *
      * @return string
      */
-    public static function ayrilmaHali($noun) {
+    public static function ayrilmaHali($noun, $fake = false) {
+        $noun = trim($noun);
+
+        //  Check the ending if it's a number
+        if(preg_match('/\d$/', $noun)) return self::processForNumber(__FUNCTION__, $noun);
+
         $last_letter      = mb_substr($noun, -1, 1, 'utf-8');
         $last_3_letter    = mb_substr($noun, -3, 3, 'utf-8');
         $last_2_letter    = mb_substr($noun, -2, 2, 'utf-8');
@@ -177,21 +195,28 @@ class Turkce {
         if(preg_match('/(*UTF8)[ouaıOUAI]/', $sh[0][count($sh[0]) - 1])) $suffix_ = "a";
         if(preg_match('/(*UTF8)[öüeiÖÜEİ]/', $sh[0][count($sh[0]) - 1])) $suffix_ = "e";
 
+        if($fake !== false) $noun = $fake;
         return ($noun . "'" . $blending . $suffix . $suffix_ . "n");
     }
 
-    public static function ablativeCase($noun) {
-        return self::ayrilmaHali($noun);
+    public static function ablativeCase($noun, $fake = false) {
+        return self::ayrilmaHali($noun, $fake);
     }
 
     /**
      * Genitive case
      *
      * @param  string $noun
+     * @param bool $fake (integer/string if it's a fake noun): this is for pronounce usage
      *
      * @return string
      */
-    public static function sahiplikHali($noun) {
+    public static function sahiplikHali($noun, $fake = false) {
+        $noun = trim($noun);
+
+        //  Check the ending if it's a number
+        if(preg_match('/\d$/', $noun)) return self::processForNumber(__FUNCTION__, $noun);
+
         $last_letter   = mb_substr($noun, -1, 1, 'utf-8');
         $last_3_letter = mb_substr($noun, -3, 3, 'utf-8');
 
@@ -204,26 +229,32 @@ class Turkce {
 
         if(count($sh[0]) == 0) return ($noun . "'" . $blending . $suffix . "n");
 
-        if(preg_match('/(*UTF8)[eiEİ]/', $sh[0][count($sh[0]) - 1])) $suffix = "i"; //  probably will be removed in next version.
         if(preg_match('/(*UTF8)[öüÖÜ]/', $sh[0][count($sh[0]) - 1])) $suffix = "ü";
         if(preg_match('/(*UTF8)[aıAI]/', $sh[0][count($sh[0]) - 1])) $suffix = "ı";
         if(preg_match('/(*UTF8)[ouOU]/', $sh[0][count($sh[0]) - 1])) $suffix = "u";
 
+        if($fake !== false) $noun = $fake;
         return ($noun . "'" . $blending . $suffix . "n");
     }
 
-    public static function genitiveCase($noun) {
-        return self::sahiplikHali($noun);
+    public static function genitiveCase($noun, $fake = false) {
+        return self::sahiplikHali($noun, $fake);
     }
 
     /**
      * Comitative case
      *
      * @param  string $noun
+     * @param bool $fake (integer/string if it's a fake noun): this is for pronounce usage
      *
      * @return string
      */
-    public static function vasitaHali($noun) {
+    public static function vasitaHali($noun, $fake = false) {
+        $noun = trim($noun);
+
+        //  Check the ending if it's a number
+        if(preg_match('/\d$/', $noun)) return self::processForNumber(__FUNCTION__, $noun);
+
         $last_letter   = mb_substr($noun, -1, 1, 'utf-8');
         $last_3_letter = mb_substr($noun, -3, 3, 'utf-8');
 
@@ -236,24 +267,30 @@ class Turkce {
 
         if(count($sh[0]) == 0) return ($noun . "'" . $blending . "l" . $suffix);
 
-        if(preg_match('/(*UTF8)[öüeiÖÜEİ]/', $sh[0][count($sh[0]) - 1])) $suffix = "e"; //  probably will be removed in next version.
         if(preg_match('/(*UTF8)[ouaıOUAI]/', $sh[0][count($sh[0]) - 1])) $suffix = "a";
 
+        if($fake !== false) $noun = $fake;
         return ($noun . "'" . $blending . "l" . $suffix);
     }
 
-    public static function comitativeCase($noun) {
-        return self::vasitaHali($noun);
+    public static function comitativeCase($noun, $fake = false) {
+        return self::vasitaHali($noun, $fake);
     }
 
     /**
      * Conjunction  (I'm not sure if this is the correct meaning of this)
      *
      * @param  string $noun
+     * @param bool $fake (integer/string if it's a fake noun): this is for pronounce usage
      *
      * @return string
      */
-    public static function dahiBaglac($noun) {
+    public static function dahiBaglac($noun, $fake = false) {
+        $noun = trim($noun);
+
+        //  Check the ending if it's a number
+        if(preg_match('/\d$/', $noun)) return self::processForNumber(__FUNCTION__, $noun);
+
         $last_3_letter = mb_substr($noun, -3, 3, 'utf-8');
         $blending      = "d";
 
@@ -263,14 +300,97 @@ class Turkce {
 
         if(count($sh[0]) == 0) return ($noun . " " . $blending . $suffix);
 
-        if(preg_match('/(*UTF8)[eiEİöüÖÜ]/', $sh[0][count($sh[0]) - 1])) $suffix = "e"; //  probably will be removed in next version.
         if(preg_match('/(*UTF8)[aıAIouOU]/', $sh[0][count($sh[0]) - 1])) $suffix = "a";
 
+        if($fake !== false) $noun = $fake;
         return ($noun . " " . $blending . $suffix);
     }
 
-    public static function conjunction($noun) {
-        return self::dahiBaglac($noun);
+    public static function conjunction($noun, $fake = false) {
+        return self::dahiBaglac($noun, $fake);
     }
+
+    /**
+     * This is for the numbers. It returns $callback function with given number's pronounce
+     *
+     * @param string $callback
+     * @param $noun
+     * @return mixed
+     */
+    public static function processForNumber($callback = 'belirtmeHali', $noun) {
+        $last_number = end(preg_split('/\D/', $noun));
+
+        if($last_number == 0) {
+            // it is directly zero
+            return call_user_func('self::' . $callback, "fır", $noun);
+        }
+
+        //  We don't need a full word, last 3 letters'd be enough
+        $endings = [
+            '1' => 'bir',
+            '2' => 'iki',
+            '3' => 'üç',
+            '4' => 'ört',
+            '5' => 'beş',
+            '6' => 'ltı',
+            '7' => 'edi',
+            '8' => 'kiz',
+            '9' => 'kuz',
+            '10' => 'on',
+            '20' => 'rmi',
+            '30' => 'tuz',
+            '40' => 'ırk',
+            '50' => 'lli',
+            '60' => 'mış',
+            '70' => 'miş',
+            '80' => 'sen',
+            '90' => 'san',
+        ];
+
+        //  last number's length
+        $digit_length = strlen($last_number);
+
+        //  Check last number
+        $last_1_letter = mb_substr($noun, -1, 1, 'utf-8');
+        $last_1_number = end(preg_split('/^\D/', $last_1_letter));
+        if($last_1_number != "0") return call_user_func('self::' . $callback, $endings[intval($last_1_number)], $noun);
+
+
+        //  Check last two numbers
+        $last_2_letter = mb_substr($noun, -2, 2, 'utf-8');
+        $last_2_number = end(preg_split('/^\D/', $last_2_letter));
+        if($last_2_number != "00") return call_user_func('self::' . $callback, $endings[intval($last_2_number)], $noun);
+
+        //  Check last three numbers
+        $last_3_letter = mb_substr($noun, -3, 3, 'utf-8');
+        $last_3_number = end(preg_split('/^\D/', $last_3_letter));
+        if($last_3_number != "000") return call_user_func('self::' . $callback, "yüz", $noun);
+
+        //  Check last four numbers
+        $last_4_letter = mb_substr($noun, -4, 4, 'utf-8');
+        $last_4_number = end(preg_split('/^\D/', $last_4_letter));
+        if($last_4_number != "0000" || $digit_length < 7) return call_user_func('self::' . $callback, "bin", $noun);
+
+        //  Check for the rest
+        if($digit_length >= 7) {
+            // if this is 8, 11, 14, etc.
+            if(($digit_length + 1) % 3 == 0) {
+                if($digit_length % 2 == 0) {
+                    return call_user_func('self::' . $callback, "yon", $noun);
+                } else {
+                    return call_user_func('self::' . $callback, "yar", $noun);
+                }
+            } elseif($digit_length % 3 == 0 || ($digit_length + 2) % 3 == 0) {
+                if($digit_length % 2 == 0) {
+                    return call_user_func('self::' . $callback, "yar", $noun);
+                } else {
+                    return call_user_func('self::' . $callback, "yon", $noun);
+                }
+            }
+        }
+
+        return call_user_func('self::' . $callback, $noun);
+    }
+
 
 }